fix reponse type

+++ JAS_API/Application/ViewModels/WalletDTOs/WalletDTO.cs
@@ -5,7 +5,7 @@ namespace Application.ViewModels.WalletDTOs
     public class WalletDTO
     {
         public int Id { get; set; }
-        public float? Balance { get; set; }
+        public Decimal? Balance { get; set; }
         public Decimal? AvailableBalance { get; set; }
         public Decimal? FrozenBalance { get; set; }
         public virtual CustomerDTO? CustomerDTO { get; set; }
